requirements.txt
fastapi
uvicorn
websockets
redis
torch
scikit-learn
pandas
joblib
twilio
pytest
numpy
prometheus_client
paho-mqtt
httpx
python-jose[cryptography]
passlib[bcrypt]
python-multipart
sqlalchemy
psycopg2-binary
alembic